Move most logic into lib.rs
Having a library at all is needed to make doctests even run
